refactor test files rename test files for consistency rename TestHelpers folder to helpers create index.js to import test files in the order they are to be run
