+++ package.json
@@ -14,10 +14,10 @@
     "test:actions": "babel-node ./node_modules/.bin/_mocha client/tests/actions/*.spec.js",
     "test:client:watch": "npm test -- --watch",
     "db:migrate": "sequelize db:migrate",
-    "seed": "NODE_ENV=development babel-node server/test/TestHelpers/SeedDatabaseData.js",
-    "reset:test": "NODE_ENV=test babel-node server/test/TestHelpers/ResetDatabase.js",
-    "seed:test": "NODE_ENV=test babel-node server/test/TestHelpers/SeedDatabaseData.js",
-    "test:server": "NODE_ENV=test babel-node ./node_modules/babel-istanbul/lib/cli.js cover ./node_modules/.bin/_mocha -- --timeout 15000 server/test/**/*.spec.js && gulp coveralls",
+    "seed": "NODE_ENV=development babel-node server/test/helpers/SeedDatabaseData.js",
+    "reset:test": "NODE_ENV=test babel-node server/test/helpers/ResetDatabase.js",
+    "seed:test": "NODE_ENV=test babel-node server/test/helpers/SeedDatabaseData.js",
+    "test:server": "NODE_ENV=test babel-node ./node_modules/babel-istanbul/lib/cli.js cover ./node_modules/.bin/_mocha -- --timeout 15000 server/test/tests/index.js && gulp coveralls",
     "build:dist": "rimraf dist/ && babel ./ --out-dir dist/ --ignore ./node_modules,./server/test,./client/test,./reports,./migrations,./coverage,docs --copy-files",
     "document": "rimraf docs/ && jsdoc -r -c ./conf.json --readme ./documentation/README.md --destination docs"
   },
