+++ server/test/Tests/Models/Role.model.spec.js
@@ -1,14 +1,13 @@
 import chai from 'chai';
 import models from '../../../models';
-import databaseData from '../../TestHelpers/DatabaseData';
+import databaseData from '../../helpers/DatabaseData';
 
 const admin = databaseData.admin;
 
 const expect = chai.expect;
 const Role = models.Role;
 
-describe('The Role model', function () {
-  this.timeout(10000);
+describe('The Role model', () => {
   describe('When supplied with invalid roles', () => {
     it('does not create duplicate roles', (done) => {
       Role.create(admin)
