+++ server/test/tests/index.js
@@ -1,0 +1,7 @@
+import './controllers/users.spec';
+import './controllers/documents.spec';
+import './controllers/roles.spec';
+import './controllers/search.spec';
+import './models/document.model.spec';
+import './models/role.model.spec';
+import './models/user.model.spec';
